+++ src/Resources/contao/config/config.php
@@ -1,4 +1,4 @@
-<?php if (!defined('TL_ROOT')) die('You cannot access this file directly!');
+<?php
 
 /**
  * Contao Open Source CMS
@@ -11,11 +11,27 @@
  * @copyright Frank Hoppe 2014
  */
 
+// Standard-CSS einbinden
+
+if(TL_MODE == 'BE') 
+{
+	$GLOBALS['TL_CSS'][] = 'bundles/contaoforum/css/be.css';
+}
+
+if(TL_MODE == 'FE') 
+{
+	$GLOBALS['TL_CSS'][] = 'bundles/contaoforum/css/style.css'; 
+	$GLOBALS['TL_CSS'][] = 'bundles/contaoforum/js/upload2.css'; 
+	$GLOBALS['TL_HEAD'][] = '<script src="bundles/contaoforum/tinymce/tiny_mce.js"></script>';
+	$GLOBALS['TL_HEAD'][] = '<script src="bundles/contaoforum/tinymce/tinymce_config.js"></script>'; 
+	$GLOBALS['TL_HEAD'][] = '<script src="bundles/contaoforum/js/upload2.js"></script>'; 
+}
+
 /**
  * Backend-Module
  */
 
-$GLOBALS['BE_MOD']['content']['forum'] = array
+$GLOBALS['BE_MOD']['forum']['tl_forum_foren'] = array
 (
 	'tables'         => array('tl_forum', 'tl_forum_threads', 'tl_forum_topics'),
 	'icon'           => 'bundles/contaoforum/images/icon.png',
@@ -26,13 +42,3 @@ $GLOBALS['BE_MOD']['content']['forum'] = array
  */
 $GLOBALS['FE_MOD']['application']['forum'] = 'Schachbulle\ContaoForumBundle\Modules\Forum';  
 
-// Standard-CSS einbinden
-if(TL_MODE == 'FE') 
-{
-	$GLOBALS['TL_CSS'][] = 'bundles/contaoforum/css/style.css'; 
-	$GLOBALS['TL_CSS'][] = 'bundles/contaoforum/js/upload2.css'; 
-	$GLOBALS['TL_HEAD'][] = '<script src="bundles/contaoforum/tinymce/tiny_mce.js"></script>';
-	$GLOBALS['TL_HEAD'][] = '<script src="bundles/contaoforum/tinymce/tinymce_config.js"></script>'; 
-	$GLOBALS['TL_HEAD'][] = '<script src="bundles/contaoforum/js/upload2.js"></script>'; 
-}
-if(TL_MODE == 'BE') $GLOBALS['TL_CSS'][] = 'bundles/contaoforum/css/be.css';