+++ src/Resources/contao/dca/tl_forum_threads.php
@@ -99,10 +99,17 @@ $GLOBALS['TL_DCA']['tl_forum_threads'] = array
 			),
 			'toggle' => array
 			(
-				'label'               => &$GLOBALS['TL_LANG']['tl_forum_threads']['toggle'],
-				'icon'                => 'visible.gif',
-				'attributes'          => 'onclick="Backend.getScrollOffset();return AjaxRequest.toggleVisibility(this,%s)"',
-				'button_callback'     => array('tl_forum_threads', 'toggleIcon') 
+				'label'                => &$GLOBALS['TL_LANG']['tl_forum_threads']['toggle'],
+				'attributes'           => 'onclick="Backend.getScrollOffset()"',
+				'haste_ajax_operation' => array
+				(
+					'field'            => 'published',
+					'options'          => array
+					(
+						array('value' => '', 'icon' => 'invisible.svg'),
+						array('value' => '1', 'icon' => 'visible.svg'),
+					),
+				),
 			),
 			'show' => array
 			(
@@ -170,21 +177,31 @@ $GLOBALS['TL_DCA']['tl_forum_threads'] = array
 			'eval'                    => array('mandatory'=>true, 'maxlength'=>128, 'tl_class'=>'long'),
 			'sql'                     => "varchar(128) NOT NULL default ''"
 		),
-		// Name des Threaderstellers (Falls leer, dann Gast)
-		'name' => array
+		// Benutzer-ID des Threaderstellers (Falls leer, dann Gast)
+		'userid' => array
 		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_forum_threads']['name'],
+			'label'                   => &$GLOBALS['TL_LANG']['tl_forum_threads']['userid'],
 			'exclude'                 => true,
-			'inputType'               => 'select',
-			'foreignKey'              => 'tl_member.username',
+			'inputType'               => 'text',
 			'eval'                    => array
 			(
 				'mandatory'           => false,
 				'tl_class'            => 'w50',
-				'choosen'             => true,
-				'includeBlankOption'  => true,
 			),
 			'sql'                     => "int(10) unsigned NOT NULL default '0'"
+		), 
+		// Benutzer-Name des Threaderstellers (Falls leer, dann Gast)
+		'username' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_forum_threads']['username'],
+			'exclude'                 => true,
+			'inputType'               => 'text',
+			'eval'                    => array
+			(
+				'mandatory'           => false,
+				'tl_class'            => 'w50',
+			),
+			'sql'                     => "varchar(64) BINARY NULL"
 		), 
 		// Email des Threaderstellers
 		'email' => array
@@ -289,74 +306,6 @@ class tl_forum_threads extends Backend
 		
 		return $line;
 	
-	}
-	
-	/**
-	 * Return the "toggle visibility" button
-	 * @param array
-	 * @param string
-	 * @param string
-	 * @param string
-	 * @param string
-	 * @param string
-	 * @return string
-	 */
-	public function toggleIcon($row, $href, $label, $title, $icon, $attributes)
-	{
-		$this->import('BackendUser', 'User');
-		
-		if (strlen($this->Input->get('tid')))
-		{
-			$this->toggleVisibility($this->Input->get('tid'), ($this->Input->get('state') == 0));
-			$this->redirect($this->getReferer());
-		}
-		
-		// Check permissions AFTER checking the tid, so hacking attempts are logged
-		if (!$this->User->isAdmin && !$this->User->hasAccess('tl_forum_threads::published', 'alexf'))
-		{
-			return '';
-		}
-		
-		$href .= '&amp;id='.$this->Input->get('id').'&amp;tid='.$row['id'].'&amp;state='.$row[''];
-		
-		if (!$row['published'])
-		{
-			$icon = 'invisible.gif';
-		}
-		
-		return '<a href="'.$this->addToUrl($href).'" title="'.specialchars($title).'"'.$attributes.'>'.$this->generateImage($icon, $label).'</a> ';
-	}
-	
-	/**
-	 * Disable/enable a user group
-	 * @param integer
-	 * @param boolean
-	 */
-	public function toggleVisibility($intId, $blnPublished)
-	{
-		// Check permissions to publish
-		if (!$this->User->isAdmin && !$this->User->hasAccess('tl_forum_threads::published', 'alexf'))
-		{
-			$this->log('Not enough permissions to show/hide record ID "'.$intId.'"', 'tl_forum_threads toggleVisibility', TL_ERROR);
-			$this->redirect('contao/main.php?act=error');
-		}
-	
-		$this->createInitialVersion('tl_forum_threads', $intId);
-	
-		// Trigger the save_callback
-		if (is_array($GLOBALS['TL_DCA']['tl_forum_threads']['fields']['published']['save_callback']))
-		{
-			foreach ($GLOBALS['TL_DCA']['tl_forum_threads']['fields']['published']['save_callback'] as $callback)
-			{
-				$this->import($callback[0]);
-				$blnPublished = $this->$callback[0]->$callback[1]($blnPublished, $this);
-			}
-		}
-	
-		// Update the database
-		$this->Database->prepare("UPDATE tl_forum_threads SET tstamp=". time() .", published='" . ($blnPublished ? '' : '1') . "' WHERE id=?")
-			->execute($intId);
-		$this->createNewVersion('tl_forum_threads', $intId);
 	}
 	
 }