+++ src/Resources/contao/dca/tl_forum_topics.php
@@ -91,10 +91,17 @@ $GLOBALS['TL_DCA']['tl_forum_topics'] = array
 			),
 			'toggle' => array
 			(
-				'label'               => &$GLOBALS['TL_LANG']['tl_forum_topics']['toggle'],
-				'icon'                => 'visible.gif',
-				'attributes'          => 'onclick="Backend.getScrollOffset();return AjaxRequest.toggleVisibility(this,%s)"',
-				'button_callback'     => array('tl_forum_topics', 'toggleIcon')
+				'label'                => &$GLOBALS['TL_LANG']['tl_forum_topics']['toggle'],
+				'attributes'           => 'onclick="Backend.getScrollOffset()"',
+				'haste_ajax_operation' => array
+				(
+					'field'            => 'published',
+					'options'          => array
+					(
+						array('value' => '', 'icon' => 'invisible.svg'),
+						array('value' => '1', 'icon' => 'visible.svg'),
+					),
+				),
 			),
 			'show' => array
 			(
@@ -153,22 +160,32 @@ $GLOBALS['TL_DCA']['tl_forum_topics'] = array
 			'eval'                    => array('rte'=>'tinyMCE', 'tl_class'=>'clr'),
 			'sql'                     => "mediumtext NULL"
 		),
-		// Name des Antworterstellers
-		'name' => array
+		// Benutzer-ID des Beitragserstellers (Falls leer, dann Gast)
+		'userid' => array
 		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_forum_topics']['name'],
+			'label'                   => &$GLOBALS['TL_LANG']['tl_forum_topics']['userid'],
 			'exclude'                 => true,
-			'inputType'               => 'select',
-			'foreignKey'              => 'tl_member.username',
+			'inputType'               => 'text',
 			'eval'                    => array
 			(
 				'mandatory'           => false,
 				'tl_class'            => 'w50',
-				'choosen'             => true,
-				'includeBlankOption'  => true,
 			),
 			'sql'                     => "int(10) unsigned NOT NULL default '0'"
-		),
+		), 
+		// Benutzer-Name des Beitragserstellers (Falls leer, dann Gast)
+		'username' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_forum_topics']['username'],
+			'exclude'                 => true,
+			'inputType'               => 'text',
+			'eval'                    => array
+			(
+				'mandatory'           => false,
+				'tl_class'            => 'w50',
+			),
+			'sql'                     => "varchar(64) BINARY NULL"
+		), 
 		// Email des Antworterstellers
 		'email' => array
 		(
@@ -266,74 +283,6 @@ class tl_forum_topics extends Backend
 		
 		return $line;
 	
-	}
-	
-	/**
-	 * Return the "toggle visibility" button
-	 * @param array
-	 * @param string
-	 * @param string
-	 * @param string
-	 * @param string
-	 * @param string
-	 * @return string
-	 */
-	public function toggleIcon($row, $href, $label, $title, $icon, $attributes)
-	{
-		$this->import('BackendUser', 'User');
-		
-		if (strlen($this->Input->get('tid')))
-		{
-			$this->toggleVisibility($this->Input->get('tid'), ($this->Input->get('state') == 0));
-			$this->redirect($this->getReferer());
-		}
-		
-		// Check permissions AFTER checking the tid, so hacking attempts are logged
-		if (!$this->User->isAdmin && !$this->User->hasAccess('tl_forum_topics::published', 'alexf'))
-		{
-			return '';
-		}
-		
-		$href .= '&amp;id='.$this->Input->get('id').'&amp;tid='.$row['id'].'&amp;state='.$row[''];
-		
-		if (!$row['published'])
-		{
-			$icon = 'invisible.gif';
-		}
-		
-		return '<a href="'.$this->addToUrl($href).'" title="'.specialchars($title).'"'.$attributes.'>'.$this->generateImage($icon, $label).'</a> ';
-	}
-	
-	/**
-	 * Disable/enable a user group
-	 * @param integer
-	 * @param boolean
-	 */
-	public function toggleVisibility($intId, $blnPublished)
-	{
-		// Check permissions to publish
-		if (!$this->User->isAdmin && !$this->User->hasAccess('tl_forum_topics::published', 'alexf'))
-		{
-			$this->log('Not enough permissions to show/hide record ID "'.$intId.'"', 'tl_forum_topics toggleVisibility', TL_ERROR);
-			$this->redirect('contao/main.php?act=error');
-		}
-	
-		$this->createInitialVersion('tl_forum_topics', $intId);
-	
-		// Trigger the save_callback
-		if (is_array($GLOBALS['TL_DCA']['tl_forum_topics']['fields']['published']['save_callback']))
-		{
-			foreach ($GLOBALS['TL_DCA']['tl_forum_topics']['fields']['published']['save_callback'] as $callback)
-			{
-				$this->import($callback[0]);
-				$blnPublished = $this->$callback[0]->$callback[1]($blnPublished, $this);
-			}
-		}
-	
-		// Update the database
-		$this->Database->prepare("UPDATE tl_forum_topics SET tstamp=". time() .", published='" . ($blnPublished ? '' : '1') . "' WHERE id=?")
-			->execute($intId);
-		$this->createNewVersion('tl_forum_topics', $intId);
 	}
 	
 	/**